Point people to #symfony instead of #symfony-dev
Many people end up coming to #symfony-dev asking basic usage questions for FoS related stuffs. Maybe that is to be expected, but sending them to #symfony instead might be more appropriate both because they are more likely to get quicker answers and because #symfony-dev is supposed to be more about core Symfony development.

There might be other FoS material pointing people to #symfony-dev, but this was the obvious one I know about. I figured I could just suggest this by way of a GitHub edit PR. If you're happy the way it is that is fine, too. No harm done. :)